fix helper wizard styles
