+++ src/components/common/layout/Wrapper.js
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react'
 import { View } from 'react-native'
- import { isMobileOnly } from 'mobile-device-detect'
+import { isMobileOnly } from 'mobile-device-detect'
 import { withStyles } from '../../../lib/styles'
 import SimpleStore from '../../../lib/undux/SimpleStore'
