+++ src/components/dashboard/FaceRecognition/ZoomCapture.js
@@ -1,28 +1,12 @@
 // @flow
-import React, { useState } from 'react'
-import { Image, StyleSheet, View } from 'react-native'
-import { Text } from 'react-native-paper'
-import { isMobile } from 'mobile-device-detect'
+import React from 'react'
+import { View } from 'react-native'
 import SimpleStore from '../../../lib/undux/SimpleStore'
-import { CustomButton } from '../../common'
-
-// import { Section } from '../../common'
 import logger from '../../../lib/logger/pino-logger'
-import WebcamGood from '../../../assets/zoom/webcam_good_ok.png'
-import WebcamBad from '../../../assets/zoom/webcam_bad_ok.png'
-import MobileAngleGood from '../../../assets/zoom/zoom-face-guy-angle-good-phone.png'
-
-// import MobileAngleBad from '../../../assets/zoom/zoom-face-guy-angle-bad-phone.png'
-import WebAngleGood from '../../../assets/zoom/zoom-face-guy-angle-good-web.png'
-
-// import WebAngleOk from '../../../assets/zoom/zoom-face-guy-angle-ok-web.png'
-// import WebAngleBad from '../../../assets/zoom/zoom-face-guy-angle-bad-web.png'
-import LightingBad1 from '../../../assets/zoom/zoom-face-guy-lighting-back-web.png'
-import LightingBad2 from '../../../assets/zoom/zoom-face-guy-lighting-side-web.png'
-import LightingGood from '../../../assets/zoom/zoom-face-guy-lighting-good-web.png'
-
+import { withStyles } from '../../../lib/styles'
 import { Camera, getResponsiveVideoDimensions } from './Camera.web'
 import Zoom, { type ZoomCaptureResult } from './Zoom'
+import HelperWizard from './HelperWizard'
 
 const log = logger.child({ from: 'ZoomCapture' })
 
@@ -35,86 +19,6 @@ type ZoomCaptureProps = {
   onCaptureResult: (captureResult?: ZoomCaptureResult) => void,
   onError: (error: string) => void,
   showHelper: boolean,
-}
-
-const HelperWizard = props => {
-  const { done, skip } = props
-  const [step, setStep] = useState(0)
-  const nextStep = () => setStep(step + 1)
-  if (skip) {
-    return null
-  }
-  let text, imgs
-  switch (step) {
-    case 0:
-      if (isMobile) {
-        nextStep()
-      } else {
-        text = 'Center your webcam'
-        imgs = (
-          <View style={styles.imageViewWebCam}>
-            <Image source={WebcamBad} resizeMode={'contain'} style={styles.webcamImage} />
-            <Image source={WebcamGood} resizeMode={'contain'} style={styles.webcamImage} />
-          </View>
-        )
-      }
-      break
-    case 1:
-      text = 'Ensure camera is at eye level'
-      if (isMobile) {
-        imgs = (
-          <View style={styles.imageView}>
-            <Image source={MobileAngleGood} resizeMode={'contain'} style={styles.mobileAngleImage} />
-            {/* <Image source={MobileAngleBad} resizeMode={'contain'} style={{ width: '100%', height: 75 }} /> */}
-          </View>
-        )
-      } else {
-        imgs = (
-          <View style={styles.imageView}>
-            <Image source={WebAngleGood} resizeMode={'contain'} style={styles.mobileAngleImage} />
-            {/* <Image source={WebAngleOk} resizeMode={'contain'} style={{ width: '100%', height: 75 }} />
-            <Image source={WebAngleBad} resizeMode={'contain'} style={{ width: '100%', height: 75 }} /> */}
-          </View>
-        )
-      }
-      break
-    case 2:
-      text = 'Light your face evenly'
-      imgs = (
-        <View style={styles.imageView}>
-          <Image source={LightingBad2} resizeMode={'contain'} style={styles.loghtingBadImage} />
-          <Image source={LightingBad1} resizeMode={'contain'} style={styles.loghtingBadImage} />
-          <Image source={LightingGood} resizeMode={'contain'} style={styles.loghtingBadImage} />
-        </View>
-      )
-      break
-    case 3:
-      done()
-      break
-  }
-  if (step === 3) {
-    return null
-  }
-  return (
-    <React.Fragment>
-      <View id="background" style={styles.background} />
-      <View style={{ zIndex: 10, justifyContent: 'space-evenly', height: '100%' }}>
-        <Text fontWeight="medium" fontSize={20} color="surface">
-          {text}
-        </Text>
-        {imgs}
-        <CustomButton
-          style={{ borderColor: 'white', borderWidth: 2 }}
-          mode={'outlined'}
-          dark={true}
-          onPress={nextStep}
-          textStyle={{ color: 'white' }}
-        >
-          OK
-        </CustomButton>
-      </View>
-    </React.Fragment>
-  )
 }
 
 /**
@@ -132,24 +36,25 @@ class ZoomCapture extends React.Component<ZoomCaptureProps> {
   }
 
   cameraReady = async (track: MediaStreamTrack) => {
+    const { store, loadedZoom, showHelper, onError } = this.props
     log.debug('camera ready')
     this.videoTrack = track
     try {
       log.debug('zoom initializes capture..')
-      let zoomSDK = this.props.loadedZoom
-      this.zoom = new Zoom(zoomSDK)
+      this.zoom = new Zoom(loadedZoom)
       await this.zoom.ready
-      this.setState({ cameraReady: true }, () => this.props.store.set('loadingIndicator')({ loading: false }))
-      if (this.props.showHelper === false) {
+      this.setState({ cameraReady: true }, () => store.set('loadingIndicator')({ loading: false }))
+      if (showHelper === false) {
         this.captureUserMediaZoom()
       }
     } catch (e) {
       log.error('Failed on capture, error:', e.message, e)
-      this.props.onError(e)
+      onError(e)
     }
   }
 
   captureUserMediaZoom = async () => {
+    const { onCaptureResult, onError } = this.props
     let captureOutcome: ZoomCaptureResult
     try {
       log.debug('zoom performs capture..')
@@ -157,43 +62,46 @@ class ZoomCapture extends React.Component<ZoomCaptureProps> {
       captureOutcome = await this.zoom.capture(this.videoTrack) // TODO: handle capture errors.
       log.info({ captureOutcome })
       if (captureOutcome) {
-        this.props.onCaptureResult(captureOutcome)
+        onCaptureResult(captureOutcome)
       }
     } catch (e) {
       log.error('Failed on capture, error:', e.message, e)
-      this.props.onError(e)
+      onError(e)
     }
   }
 
   componentDidMount() {
-    this.props.store.set('loadingIndicator')({ loading: true })
-    if (!this.props.loadedZoom) {
+    const { store, loadedZoom } = this.props
+    store.set('loadingIndicator')({ loading: true })
+    if (!loadedZoom) {
       log.warn('zoomSDK was not loaded into ZoomCapture properly')
     }
   }
 
   componentWillUnmount() {
-    if (this.state.cameraReady === false) {
-      this.props.store.set('loadingIndicator')({ loading: false })
+    const { cameraReady } = this.state
+    const { loadedZoom, store } = this.props
+    if (cameraReady === false) {
+      store.set('loadingIndicator')({ loading: false })
     }
-    if (this.props.loadedZoom) {
+    if (loadedZoom) {
       log.warn('zoomSDK was loaded, canceling zoom capture')
       this.zoom && this.zoom.cancel()
     }
   }
 
   render() {
+    const { styles, showHelper, onError } = this.props
+    const { cameraReady } = this.state
     return (
       <View>
         <View style={styles.bottomSection}>
           <div id="zoom-parent-container" style={getVideoContainerStyles()}>
             <View id="helper" style={styles.helper}>
-              {this.state.cameraReady ? (
-                <HelperWizard done={this.captureUserMediaZoom} skip={this.props.showHelper === false} />
-              ) : null}
+              {cameraReady ? <HelperWizard done={this.captureUserMediaZoom} skip={showHelper === false} /> : null}
             </View>
-            <div id="zoom-interface-container" style={{ position: 'absolute' }} />
-            {<Camera key="camera" onCameraLoad={this.cameraReady} onError={this.props.onError} />}
+            <div id="zoom-interface-container" style={zoomInterfaceStyle} />
+            {<Camera key="camera" onCameraLoad={this.cameraReady} onError={onError} />}
           </div>
         </View>
       </View>
@@ -201,35 +109,13 @@ class ZoomCapture extends React.Component<ZoomCaptureProps> {
   }
 }
 
-const styles = StyleSheet.create({
-  imageViewWebCam: {
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  imageView: {
-    display: 'flex',
-    flexDirection: 'column',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  webcamImage: {
-    margin: '0.9rem',
-    width: '100%',
-    height: '7.5rem',
-  },
-  loghtingBadImage: {
-    width: '100%',
-    height: '4.6875rem',
-  },
-  mobileAngleImage: {
-    width: '100%',
-    height: '4.6875rem',
+const getStylesFromProps = ({ theme }) => ({
+  zoomInterfaceContainer: {
+    position: 'absolute',
   },
   bottomSection: {
     flex: 1,
-    backgroundColor: '#fff',
+    backgroundColor: theme.colors.white,
     padding: 5,
     borderRadius: 5,
   },
@@ -243,27 +129,10 @@ const styles = StyleSheet.create({
     justifyContent: 'space-around',
     alignItems: 'center',
   },
-  background: {
-    position: 'absolute',
-    top: 0,
-    left: -5,
-    zIndex: 8,
-    borderWidth: 5,
-    borderColor: 'white',
-    width: '100%',
-    height: '100%',
-    backgroundColor: 'rgba(13, 165, 177, 0.5)',
-    borderRadius: 5,
-
-    // background:
-    //   'linear-gradient(to right, rgba(9, 181, 163, .5), rgba(18, 146, 193, .95)) no-repeat center center fixed',
-    backgroundImage: 'linear-gradient(to right, rgba(9, 181, 163, .5), rgba(18, 146, 193, .95))',
-    backgroundRepeat: 'no-repeat',
-    backgroundPosition: 'center center',
-    backgroundAttachment: 'fixed',
-  },
 })
-
+const zoomInterfaceStyle = {
+  position: 'absolute',
+}
 const getVideoContainerStyles = () => ({
   ...getResponsiveVideoDimensions(),
   marginLeft: '0',
@@ -272,4 +141,4 @@ const getVideoContainerStyles = () => ({
   marginBottom: 0,
 })
 
-export default SimpleStore.withStore(ZoomCapture)
+export default SimpleStore.withStore(withStyles(getStylesFromProps)(ZoomCapture))
